# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: AI assistant operations
  - name: billing
    description: Pass purchase and entitlement operations
  - name: tools
    description: Pregnancy tracking tools
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/ask:
    post:
      operationId: askAssistant
      tags: [ai]
      summary: Ask the pregnancy assistant
      description: >
        Ask an evidence-based pregnancy and infant-health question. Answers are
        guided to align with ACOG and AAP guidance and always include a
        disclaimer to consult a healthcare provider.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQuery"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Free question limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many requests; rate limit exceeded
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: The AI provider failed to respond
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The AI integration is not provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ai/weekly-insight:
    post:
      operationId: getWeeklyInsight
      tags: [ai]
      summary: Weekly pregnancy insight
      description: >
        Returns an evidence-based insight for a given gestational week, guided to
        align with ACOG and AAP guidance, with a disclaimer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeeklyInsightQuery"
      responses:
        "200":
          description: Weekly insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyInsight"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many requests; rate limit exceeded
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: The AI provider failed to respond
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The AI integration is not provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/pass-status:
    get:
      operationId: getPassStatus
      tags: [billing]
      summary: Pass entitlement and free usage
      description: >
        Returns whether the signed-in user owns the Full Pregnancy Pass and
        their weekly free question usage.
      responses:
        "200":
          description: Pass status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassStatus"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Start a checkout session for the pass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Payments unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/confirm:
    post:
      operationId: confirmCheckout
      tags: [billing]
      summary: Confirm a completed checkout and refresh entitlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Updated pass status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassStatus"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/revenuecat/reconcile:
    post:
      operationId: reconcileRevenuecat
      tags: [billing]
      summary: Reconcile a RevenueCat in-app purchase and refresh entitlement
      description: >
        Verifies the signed-in user's active RevenueCat entitlement (the Clerk
        user id is used as the RevenueCat app user id) and grants the Full
        Pregnancy Pass when the entitlement is active. The pass is recorded on
        the same backend account, so it also unlocks unlimited AI on web.
      responses:
        "200":
          description: Updated pass status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassStatus"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Purchases are not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/redeem-code:
    post:
      operationId: redeemCode
      tags: [billing]
      summary: Redeem a developer access code to unlock the pass
      description: >
        Grants the Full Pregnancy Pass to the signed-in user when the provided
        code matches the server-configured developer access code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemCodeRequest"
      responses:
        "200":
          description: Updated pass status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassStatus"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Invalid access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Code redemption is not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/symptoms:
    get:
      operationId: listSymptoms
      tags: [tools]
      summary: List symptom logs for the signed-in user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Symptom log list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymptomLogList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createSymptom
      tags: [tools]
      summary: Log a new symptom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSymptomBody"
      responses:
        "201":
          description: Created symptom log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymptomLog"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/symptoms/{id}:
    delete:
      operationId: deleteSymptom
      tags: [tools]
      summary: Delete a symptom log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/kick-sessions:
    get:
      operationId: listKickSessions
      tags: [tools]
      summary: List kick counter sessions for the signed-in user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Kick session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KickSessionList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createKickSession
      tags: [tools]
      summary: Create a new kick counter session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKickSessionBody"
      responses:
        "201":
          description: Created kick session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KickSession"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/kick-sessions/{id}:
    patch:
      operationId: updateKickSession
      tags: [tools]
      summary: Update a kick counter session (add kicks, end session)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKickSessionBody"
      responses:
        "200":
          description: Updated kick session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KickSession"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteKickSession
      tags: [tools]
      summary: Delete a kick session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/contractions:
    get:
      operationId: listContractions
      tags: [tools]
      summary: List contraction logs for the signed-in user
      parameters:
        - name: sessionDate
          in: query
          description: Filter by session date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: Contraction log list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractionList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createContraction
      tags: [tools]
      summary: Log a completed contraction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractionBody"
      responses:
        "201":
          description: Created contraction log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contraction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/contractions/{id}:
    delete:
      operationId: deleteContraction
      tags: [tools]
      summary: Delete a contraction log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
    PassStatus:
      type: object
      properties:
        hasPass:
          type: boolean
        freeUsed:
          type: integer
        freeLimit:
          type: integer
      required:
        - hasPass
        - freeUsed
        - freeLimit
    CheckoutRequest:
      type: object
      properties:
        returnUrl:
          type: string
          minLength: 1
      required:
        - returnUrl
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    ConfirmRequest:
      type: object
      properties:
        sessionId:
          type: ["string", "null"]
    RedeemCodeRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
      required:
        - code
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
      required:
        - role
        - content
    AssistantQuery:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
        week:
          type: ["integer", "null"]
          minimum: 1
          maximum: 42
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - question
    AssistantReply:
      type: object
      properties:
        answer:
          type: string
        disclaimer:
          type: string
      required:
        - answer
        - disclaimer
    WeeklyInsightQuery:
      type: object
      properties:
        week:
          type: integer
          minimum: 1
          maximum: 42
      required:
        - week
    WeeklyInsight:
      type: object
      properties:
        week:
          type: integer
        insight:
          type: string
        disclaimer:
          type: string
      required:
        - week
        - insight
        - disclaimer
    SymptomLog:
      type: object
      properties:
        id:
          type: integer
        symptom:
          type: string
        severity:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        loggedAt:
          type: string
          format: date-time
      required:
        - id
        - symptom
        - loggedAt
    SymptomLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SymptomLog"
        total:
          type: integer
      required:
        - items
        - total
    CreateSymptomBody:
      type: object
      properties:
        symptom:
          type: string
          minLength: 1
          maxLength: 200
        severity:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        notes:
          type: ["string", "null"]
          maxLength: 1000
        loggedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - symptom
    KickSession:
      type: object
      properties:
        id:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        kickCount:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - id
        - startedAt
        - kickCount
    KickSessionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KickSession"
        total:
          type: integer
      required:
        - items
        - total
    CreateKickSessionBody:
      type: object
      properties:
        startedAt:
          type: ["string", "null"]
          format: date-time
        kickCount:
          type: ["integer", "null"]
          minimum: 0
        endedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
          maxLength: 500
    UpdateKickSessionBody:
      type: object
      properties:
        kickCount:
          type: ["integer", "null"]
          minimum: 0
        endedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
          maxLength: 500
    Contraction:
      type: object
      properties:
        id:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
        intervalSeconds:
          type: ["integer", "null"]
        sessionDate:
          type: string
      required:
        - id
        - startedAt
        - sessionDate
    ContractionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Contraction"
        total:
          type: integer
      required:
        - items
        - total
    CreateContractionBody:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
          minimum: 0
        intervalSeconds:
          type: ["integer", "null"]
          minimum: 0
        sessionDate:
          type: string
          minLength: 1
      required:
        - startedAt
        - sessionDate