# lib/api-zod/src/generated/types/contraction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Contraction {
  id: number;
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  intervalSeconds?: number | null;
  sessionDate: string;
}