# lib/api-zod/src/generated/types/contractionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contraction } from "./contraction";

export interface ContractionList {
  items: Contraction[];
  total: number;
}