# lib/api-zod/src/generated/types/createContractionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractionBody {
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  /**
   * @minimum 0
   * @nullable
   */
  durationSeconds?: number | null;
  /**
   * @minimum 0
   * @nullable
   */
  intervalSeconds?: number | null;
  /** @minLength 1 */
  sessionDate: string;
}