# lib/api-zod/src/generated/types/createKickSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateKickSessionBody {
  /** @nullable */
  startedAt?: Date | null;
  /**
   * @minimum 0
   * @nullable
   */
  kickCount?: number | null;
  /** @nullable */
  endedAt?: Date | null;
  /**
   * @maxLength 500
   * @nullable
   */
  notes?: string | null;
}