# lib/api-zod/src/generated/types/createSymptomBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSymptomBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  symptom: string;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  severity?: number | null;
  /**
   * @maxLength 1000
   * @nullable
   */
  notes?: string | null;
  /** @nullable */
  loggedAt?: Date | null;
}