# lib/api-zod/src/generated/types/kickSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KickSession {
  id: number;
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  kickCount: number;
  /** @nullable */
  notes?: string | null;
}