# lib/api-zod/src/generated/types/kickSessionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KickSession } from "./kickSession";

export interface KickSessionList {
  items: KickSession[];
  total: number;
}