# lib/api-zod/src/generated/types/symptomLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SymptomLog {
  id: number;
  symptom: string;
  /** @nullable */
  severity?: number | null;
  /** @nullable */
  notes?: string | null;
  loggedAt: Date;
}